Split simple ops into its own file.
